Rename RaspberryPi3Driver to common base class GenericRaspberryPiDriver
