+++ src/System.Device.Gpio/System/Device/Gpio/Drivers/RaspberryPi3Driver.cs
@@ -7,60 +7,8 @@ namespace System.Device.Gpio.Drivers
     /// <summary>
     /// A GPIO driver for the Raspberry Pi 3.
     /// </summary>
-    public partial class RaspberryPi3Driver  // Different base classes declared in RaspberryPi3Driver.Linux.cs and RaspberryPi3Driver.Windows.cs
+    public partial class RaspberryPi3Driver : GenericRaspberryPiDriver
     {
-        /// <summary>
-        /// Raspberry Pi 3 has 28 GPIO pins.
-        /// </summary>
-        protected internal override int PinCount => 28;
-
-        protected void ValidatePinNumber(int pinNumber)
-        {
-            if (pinNumber < 0 || pinNumber > 27)
-            {
-                throw new ArgumentException("The specified pin number is invalid.", nameof(pinNumber));
-            }
-        }
-
-        /// <summary>
-        /// Converts a board pin number to the driver's logical numbering scheme.
-        /// </summary>
-        /// <param name="pinNumber">The board pin number to convert.</param>
-        /// <returns>The pin number in the driver's logical numbering scheme.</returns>
-        protected internal override int ConvertPinNumberToLogicalNumberingScheme(int pinNumber)
-        {
-            return pinNumber switch
-            {
-                3 => 2,
-                5 => 3,
-                7 => 4,
-                8 => 14,
-                10 => 15,
-                11 => 17,
-                12 => 18,
-                13 => 27,
-                15 => 22,
-                16 => 23,
-                18 => 24,
-                19 => 10,
-                21 => 9,
-                22 => 25,
-                23 => 11,
-                24 => 8,
-                26 => 7,
-                27 => 0,
-                28 => 1,
-                29 => 5,
-                31 => 6,
-                32 => 12,
-                33 => 13,
-                35 => 19,
-                36 => 16,
-                37 => 26,
-                38 => 20,
-                40 => 21,
-                _ => throw new ArgumentException($"Board (header) pin {pinNumber} is not a GPIO pin on the {GetType().Name} device.", nameof(pinNumber))
-            };
-        }
+        // No extra code required for Pi3 only for now
     }
 }
